student_support_adk/requirements.txt
flask
google-genai
google-adk
requests
python-dotenv
pytest
